Fix: virtual calls.
Certain virtual calls would result in generating a model which could not compile. (1) issue was that super calls only change the scope of the single call. (2) issue was that super calls can skip implementations in the linearized inheritance hierarchy.

Before this diff, we were using an ad-hoc, 'call-graph-like' analysis. The analysis was not consistent, and inconsistencies lead to these bugs in edge cases. This diff makes the call graph explicit, and propogates it throughout the analysis.